+++ src/hyp3_autorift/__main__.py
@@ -2,7 +2,6 @@
 AutoRIFT processing for HyP3
 """
 
-
 import argparse
 import os
 import sys
@@ -16,8 +15,10 @@ from hyp3lib.fetch import write_credentials_to_netrc_file
 def main():
     parser = argparse.ArgumentParser(prefix_chars='+', formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     parser.add_argument(
-        '++process', choices=['hyp3_autorift', 's1_correction'], default='hyp3_autorift',
-        help='Select the console_script entrypoint to use'  # console_script entrypoints are specified in `setup.py`
+        '++process',
+        choices=['hyp3_autorift', 's1_correction'],
+        default='hyp3_autorift',
+        help='Select the console_script entrypoint to use',  # console_script entrypoints are specified in `setup.py`
     )
     parser.add_argument('++omp-num-threads', type=int, help='The number of OpenMP threads to use for parallel regions')
 
@@ -38,9 +39,7 @@ def main():
     (process_entry_point,) = {process for process in eps if process.name == args.process}
 
     sys.argv = [args.process, *unknowns]
-    sys.exit(
-        process_entry_point.load()()
-    )
+    sys.exit(process_entry_point.load()())
 
 
 if __name__ == '__main__':